requirements.txt
Django==2.1.4
django-phone-field==1.6.0
pytz==2018.7
